Travail sur l'agencement des blocs
