+++ source/components/Simu.css
@@ -1,0 +1,3 @@
+#focusZone {
+		min-height: 100%
+}
