+++ source/components/conversation/Conversation.js
@@ -2,18 +2,10 @@ import React, { Component } from 'react'
 import { isEmpty, map } from 'ramda'
 import Aide from '../Aide'
 import { reduxForm, reset } from 'redux-form'
-import { scroller, Element } from 'react-scroll'
 import { getInputComponent } from 'Engine/generateQuestions'
 import Satisfaction from '../Satisfaction'
 import { connect } from 'react-redux'
 import './conversation.css'
-
-let scroll = () =>
-	scroller.scrollTo('myScrollToElement', {
-		duration: 500,
-		delay: 0,
-		smooth: true
-	})
 
 @reduxForm({
 	form: 'conversation',
@@ -40,12 +32,6 @@ let scroll = () =>
 	})
 )
 export default class Conversation extends Component {
-	componentWillReceiveProps(nextProps) {
-		if (nextProps.foldedSteps.length == this.props.foldedSteps.length)
-			return null
-
-		setTimeout(scroll, 1) // scrolling after the first answer doesn't work
-	}
 	render() {
 		let {
 			foldedSteps,
@@ -57,18 +43,16 @@ export default class Conversation extends Component {
 		} = this.props
 		return (
 			<>
-				<Element name="myScrollToElement" id="myScrollToElement">
-					<div id="currentQuestion">
-						{currentQuestion ? (
-							getInputComponent({ unfolded: true })(parsedRules, targetNames)(
-								currentQuestion
-							)
-						) : (
-							<Satisfaction />
-						)}
-					</div>
-					<Aide />
-				</Element>
+				<div id="currentQuestion">
+					{currentQuestion ? (
+						getInputComponent({ unfolded: true })(parsedRules, targetNames)(
+							currentQuestion
+						)
+					) : (
+						<Satisfaction />
+					)}
+				</div>
+				<Aide />
 			</>
 		)
 	}
