+++ source/components/conversation/FoldedSteps.js
@@ -1,10 +1,19 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 //import styles from './.css'
+// css in conversation.Css
 import { isEmpty, map } from 'ramda'
 import ReactPiwik from '../Tracker'
 import { getInputComponent } from 'Engine/generateQuestions'
 import withColours from '../withColours'
+import { scroller, Element } from 'react-scroll'
+
+let scroll = () =>
+	scroller.scrollTo('myScrollToElement', {
+		duration: 500,
+		delay: 0,
+		smooth: true
+	})
 
 @withColours
 @connect(
@@ -50,17 +59,36 @@ export default class extends Component {
 					getInputComponent({ unfolded: false })(parsedRules, targetNames),
 					foldedSteps
 				)}
+			</div>
+		)
+	}
+}
+
+@withColours
+@connect(state => ({
+	foldedSteps: state.foldedSteps
+}))
+export class GoToAnswers extends Component {
+	componentWillReceiveProps(nextProps) {
+		if (nextProps.foldedSteps.length == this.props.foldedSteps.length)
+			return null
+
+		setTimeout(scroll, 1) // scrolling after the first answer doesn't work
+	}
+	render() {
+		return (
+			<Element name="myScrollToElement" id="myScrollToElement">
 				<h3
 					className="scrollIndication up"
 					style={{
-						opacity: foldedSteps.length != 0 ? 1 : 0,
-						color: textColourOnWhite
+						opacity: this.props.foldedSteps.length != 0 ? 1 : 0,
+						color: this.props.colours.textColourOnWhite
 					}}
 				>
 					<i className="fa fa-long-arrow-up" aria-hidden="true" /> Modifier mes
 					réponses
 				</h3>
-			</div>
+			</Element>
 		)
 	}
 }
