+++ source/containers/App.dev.js
@@ -7,7 +7,6 @@ export default ({ store }) => (
 	<Provider store={store}>
 		<div id="dev">
 			<Layout />
-			<DevTools />
 		</div>
 	</Provider>
 )
